+++ src/ansys/fluent/core/launcher/slurm_launcher.py
@@ -40,11 +40,11 @@ from ansys.fluent.core.launcher.custom_exceptions import _process_invalid_args
 from ansys.fluent.core.launcher.launcher_utils import (
     _await_fluent_launch,
     _build_journal_argument,
-    _get_server_info_file_name,
     _get_subprocess_kwargs_for_fluent,
 )
 from ansys.fluent.core.launcher.process_launch_string import _generate_launch_string
 from ansys.fluent.core.launcher.pyfluent_enums import _get_mode
+from ansys.fluent.core.launcher.server_info import _get_server_info_file_name
 from ansys.fluent.core.session_meshing import Meshing
 from ansys.fluent.core.session_pure_meshing import PureMeshing
 from ansys.fluent.core.session_solver import Solver
